Update translations and the subtitle in "berth rejected" -email, fix typos in berth_none_offered_fi.html and berth_application_confirmation{en|sv}.html
