+++ templates/email/messages/berth_application_confirmation_en.html
@@ -68,7 +68,7 @@
         </tr>
         <tr>
             <td><b>Model</b></td>
-            <td>{ application.boat_model }}</td>
+            <td>{{ application.boat_model }}</td>
         </tr>
     {% endif %}
