+++ templates/email/messages/berth_application_confirmation_sv.html
@@ -68,7 +68,7 @@
         </tr>
         <tr>
             <td><b>Modell</b></td>
-            <td>{ application.boat_model }}</td>
+            <td>{{ application.boat_model }}</td>
         </tr>
     {% endif %}
