+++ templates/email/messages/berth_none_offered_fi.html
@@ -1,4 +1,4 @@
-<p>Hyvä aiakas, valitettavasti emme pysty tarjoamaan paikkaa kohteesta jota olet hakenut.</p>
+<p>Hyvä asiakas, valitettavasti emme pysty tarjoamaan paikkaa kohteesta jota olet hakenut.</p>
 
 <h3>Kohde</h3>
